explicit cast for MacOS

+++ src/neuralNetwork.h
@@ -146,7 +146,7 @@ private:
 
     /** Parameters and functions for calculating amount of change for each weight */
     T outputErrorGradient;
-    inline T getHiddenErrorGradient(int layer, int neuron);
+    inline T getHiddenErrorGradient(int layer, size_t neuron);
 
     void initTrainer();
